Remove dynamic checkout PassKit session
